Cleaning up organization of cor_analysis code
