+++ packed-cor-tables/packed_cor_tables/__init__.py
@@ -145,58 +145,3 @@ def read_cor_for_given_id(filename, feature_id):
     conn.close()
 
     return df.drop(columns=["dim_0", "dim_1"])
-
-
-def test_write_df(tmpdir):
-    df = pd.DataFrame(
-        {
-            "dim_0": [0, 0, 1, 1],
-            "dim_1": [1, 2, 3, 0],
-            "cor": [0.1, 0.2, 0.001, 0.9],
-            "log10qvalue": [1e-10, 1e-1, 0.1, 1e-12],
-        }
-    )
-    out = str(tmpdir.join("out.sqlite3"))
-    write_cor_df(
-        df,
-        InputMatrixDesc(
-            given_ids=["C1", "C2", "C3", "C4"], taiga_id="crispr-taiga", name="crispr"
-        ),
-        InputMatrixDesc(
-            given_ids=["E1", "E2", "E3", "E4"], taiga_id="expr-taiga", name="expr"
-        ),
-        out,
-    )
-
-    df = read_cor_for_given_id(out, "C1").sort_values("feature_given_id_1")
-
-    def round_list(l):
-        return [float(f"{x:.4}") for x in l]
-
-    assert list(df["feature_given_id_1"]) == ["E2", "E3"]
-    assert round_list(df["cor"]) == [0.1, 0.2]
-
-    df = read_cor_for_given_id(out, "C2").sort_values("feature_given_id_1")
-    assert list(df["feature_given_id_1"]) == ["E1", "E4"]
-    assert round_list(df["cor"]) == [0.9, 0.001]
-
-
-# at the end of the day, we want this to return a df with columns:
-# (compound dataset) taiga_id, (compound feature) given_id, compound_name, compound_dose, dataset_name, measurement,
-# cor_taiga_id, cor_given_id, cor_label, cor_dataset_name, cor, qvalue
-
-# in the UI transform that into
-# interface CompoundCorrelation {
-#   measurement: string;
-#   dose? : number;
-#   featureDataset : string;
-#   featureLabel : string;
-#   cor :number;
-#   negLog10QValue: number;
-# }
-
-
-# Add filters for:
-#   Dose + Measurement
-#   correlated_dataset_name
-#   correlated_label
